requirements.txt
chromadb>=0.4.22
langchain>=0.1.11
langchain-community>=0.0.27
langchain-ollama>=0.1.1
langchain-core>=0.1.30
python-dotenv>=1.0.1
pypdf>=4.0.2
pydantic>=2.6.3
python-docx>=0.8.11
